First commit of the test page for class Managesieve.
You will need a server and a test account with no existing scripts. The
script will prompt first for server and login details.

If class Managesieve returns the expected values on each test, the page
will display a green Success notice, or a red Failure warning if not. Note
that some SASL method tests will certainly fail because class Managesieve
does not support all methods.

Also, the havespace tests are not ran on Cyrusv2.0 servers because the
havespace command is broken on these.